Repository cleanup + figured out how to make svg container size of parent container
