--- public/css/extra.css
@@ -1,64 +1,0 @@
-.bar {
-  fill: steelblue;
-}
-
-.bar:hover {
-  fill: brown;
-}
-
-.axis {
-  font: 10px sans-serif;
-}
-
-.axis path,
-.axis line {
-  fill: none;
-  stroke: black;
-  shape-rendering: crispEdges;
-}
-
-/*.x.axis path {
-  display: none;
-}*/
-
-.chart {
-    background: #eee;
-    padding: 3px;
-}
-
-.chart div {
-  width: 0;
-  transition: all 1s ease-out;
-  -moz-transition: all 1s ease-out;
-  -webkit-transition: all 1s ease-out;
-}
-
-.chart div {
-  font: 10px sans-serif;
-  background-color: brown;
-  text-align: right;
-  padding: 3px;
-  margin: 5px;
-  color: white;
-  box-shadow: 2px 2px 2px #666;
-}
-
-.chart div:hover {
-	background-color: steelblue;
-}
-
-
-.axis text{
-  font-family: sans-serif;
-  font-size: 10px;
-}
-
-h1{
-  font-family: sans-serif;
-  font-weight: bold;
-  font-size: 16px;
-}
-.tick
-{
-  stroke-dasharray: 1, 2;
-}
