+++ public/endurance.html
@@ -1,1 +1,3 @@
-<dir1 data="barData"></dir1>
+<div class="main-content">
+	<dir1></dir1>
+</div>
