+++ public/index.html
@@ -4,19 +4,18 @@
 <head>
   <!-- Theme CSS -->
   <link href="css/main.css" rel="stylesheet">
-  <link href="css/extra.css" rel="stylesheet">
 
   <!-- Load d3.js & color scale-->
   <!-- <script src="https://d3js.org/d3.v4.js"></script> -->
   <script TYPE="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/d3/3.5.5/d3.min.js"></script>
-  <script TYPE="text/javascript" src="https://d3js.org/d3.v3.min.js"></script>
+  <script TYPE="text/javascript" src="https://d3js.org/d3.v4.min.js"></script>
   <script src="https://d3js.org/d3-scale-chromatic.v2.min.js"></script>
 
   <!-- Routing js -->
   <script data-require="angular.js@1.4.x" src="https://code.angularjs.org/1.4.9/angular.js" data-semver="1.4.9"></script>
   <script data-require="angular.js@1.4.x" src="https://code.angularjs.org/1.4.9/angular-route.js" data-semver="1.4.9"></script>
   <script src="js/route-config.js"></script>
-  <script src="js/test.js"></script>
+  <script src="js/endurance.js"></script>
   <script src="js/headerController.js"></script>
 
   <!-- Custom Fonts -->
