--- public/test.html
@@ -1,92 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-
-    <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <meta name="description" content="">
-    <meta name="author" content="">
-
-    <title>Harvey Barnhard</title>
-
-    <!-- Theme CSS -->
-    <link href="css/main.css" rel="stylesheet">
-
-		<!-- Load d3.js & color scale-->
-		<script src="https://d3js.org/d3.v4.min.js"></script>
-		<script src="https://d3js.org/d3-scale-chromatic.v2.min.js"></script>
-
-    <!-- Custom Fonts -->
-    <script src="https://kit.fontawesome.com/b89efab99d.js" crossorigin="anonymous"></script>
-    <link href='https://fonts.googleapis.com/css?family=Lora:400,700,400italic,700italic' rel='stylesheet' type='text/css'>
-    <link href='https://fonts.googleapis.com/css?family=Open+Sans:300italic,400italic,600italic,700italic,800italic,400,300,600,700,800' rel='stylesheet' type='text/css'>
-
-    <!-- favicon -->
-    <link rel="shortcut icon" type="image/x-icon" href="img/favicon.ico">
-
-    <!-- HTML5 Shim and Respond.js IE8 support of HTML5 elements and media queries -->
-    <!-- WARNING: Respond.js doesn't work if you view the page via file:// -->
-    <!--[if lt IE 9]>
-        <script src="https://oss.maxcdn.com/libs/html5shiv/3.7.0/html5shiv.js"></script>
-        <script src="https://oss.maxcdn.com/libs/respond.js/1.4.2/respond.min.js"></script>
-    <![endif]-->
-
-</head>
-
-<body>
-
-    <!-- Navigation -->
-    <div class="navbar-container">
-      <nav class="navbar">
-        <ul>
-          <li class="about">
-            <a href="about.html"> Harvey Barnhard </a>
-          </li>
-          <li>
-            <a class="active" href="endurance.html">endurance</a>
-          </li>
-          <li>
-            <a href="blog.html">blog</a>
-          </li>
-          <li>
-            <a href="projects.html">projects</a>
-          </li>
-        </ul>
-      </nav>
-    </div>
-
-    <!-- Main Content -->
-    <div class="main-content">
-      <div id="active_time"></div>
-    </div>
-    <!-- Footer -->
-    <footer class="footer-default">
-      <ul>
-          <li>
-              <a href="https://github.com/harveybarnhard">
-                  <span class="fa-stack fa-lg">
-                      <i class="fa fa-circle fa-stack-2x"></i>
-                      <i class="fa fa-github fa-stack-1x fa-inverse"></i>
-                  </span>
-              </a>
-          </li>
-          <li>
-              <a href="https://strava.com/athletes/16562005">
-                  <span class="fa-stack fa-lg">
-                      <i class="fa fa-circle fa-stack-2x"></i>
-                      <i class="fab fa-strava fa-stack-1x fa-inverse"></i>
-                  </span>
-              </a>
-          </li>
-      </ul>
-    </footer>
-
-    <!-- jQuery -->
-    <script src="vendor/jquery/jquery.min.js"></script>
-
-</body>
-
-</html>
-<script src="js/active-hours.js"></script>
